Private options!
All options are privatised, getter/setter methods are written for all options, so that a few crash scenarios are eliminated this way.

There are two new methods for ViewPagerController:
- (CGFloat)valueForOption:(ViewPagerOption)option;
- (UIColor *)colorForComponent:(ViewPagerComponent)component;

With these methods, you can get the default or provided (by delegate methods) values for the given option or component key.