Fixed a tab scrolling bug.
Before this, sometimes tab won't follow by scrolling when content view
is paged by swipe gestures. This was caused when `animatingToTab`
property is set to `YES` and never gets back to `NO` again. So I just
worked around this by adding a completion handler to cases that was not
handling this property. I'm not sure that was intentional or not, but
for me this fix works well.