Fixed #35.
The first plan for this fix was just to workaround not to pass
`NSNotFound` to `selectTabAtIndex:`, but I found `selectTabAtIndex:`
doesn't take out-of-bounds arguments into account. In private methods
like `viewControllerAtIndex:` handles invalid arguments collectly so I
decided follow this way. Notice that `selectTabAtIndex:` is a public
method which behaviour will be changed by this fix.